fixed 4x4 board bug

+++ lib/game.h
@@ -80,7 +80,7 @@ game_t *game_init(
     node_t const root,
     size_t const width,
     int const heuristic_max,
-    uint8_t const depth,
+    unsigned const depth,
     bool player1_ai,
     bool player2_ai,
     leaf_t const leaf,
@@ -105,7 +105,7 @@ node_t game_root(game_t const *game);
 size_t game_width(game_t const *game);
 int game_heuristic_max(game_t const *game);
 negamax_t *game_negamax(game_t const *game);
-uint8_t game_depth(game_t const *game);
+unsigned game_depth(game_t const *game);
 bool game_player1_ai(game_t const *game);
 bool game_player2_ai(game_t const *game);
 player_t game_player(game_t const *game);
