+++ lib/negamax.h
@@ -17,8 +17,8 @@ void negamax_free(negamax_t *negamax);
 size_t negamax_nbytes(negamax_t const *negamax);
 size_t negamax_ttable_size(negamax_t const *negamax);
 
-int negamax_eval(negamax_t *negamax, node_t node, player_t player, uint8_t depth);
-node_t negamax_move(negamax_t *negamax, node_t node, player_t player, uint8_t depth, int * const eval);
+int negamax_eval(negamax_t *negamax, node_t node, player_t player, unsigned depth);
+node_t negamax_move(negamax_t *negamax, node_t node, player_t player, unsigned depth, int * const eval);
 //negamax_rank?
 
 #endif // NEGAMAX_H
